+++ src/constants/urn.ts
@@ -10,6 +10,7 @@
 export const UrnPrefixes = Object.freeze({
   JwkThumbprintSha256KeyId: 'urn:ietf:params:oauth:jwk-thumbprint:sha-256:',
   Multibase: 'urn:multibase:',
+  Uuid: 'urn:uuid:',
 } as const);
 
 export type UrnPrefix = typeof UrnPrefixes[keyof typeof UrnPrefixes];